Yuvish app

+++ app/src/main/java/com/example/yuvish/Models/DebtorsAPI/Market/Paydebt.kt
@@ -2,6 +2,6 @@ package com.example.yuvish.Models.DebtorsAPI.Market
 
 data class Paydebt(
     val id: Int,
-    val summa: Int,
+    val summa: Double,
     val tolov_turi: String
 )
